// apps/web/src/lib/api/generated/model/productIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { MemberTierInInput } from './memberTierInInput';
import type { ProductBenefitInput } from './productBenefitInput';
import type { ProductInRatingBasis } from './productInRatingBasis';

export interface ProductIn {
  /** @maxItems 30 */
  benefits?: ProductBenefitInput[];
  /** @pattern ^[a-z0-9_]{1,40}$ */
  class_code: string;
  commission_rate_new?: string | null;
  commission_rate_renewal?: string | null;
  /** @pattern ^[A-Z]{3}$ */
  currency?: string;
  excess_text?: string | null;
  flat_premium?: string | null;
  insurer_id: string;
  /** @maxItems 10 */
  member_tiers?: MemberTierInInput[];
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  min_premium?: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  notes?: string | null;
  rate?: string | null;
  rating_basis: ProductInRatingBasis;
}